+++ matflow_damask/main.py
@@ -427,6 +427,23 @@ def modify_volume_element_geometry(volume_element, volume_element_response):
     return out
 
 
+@func_mapper(task='modify_volume_element', method='grid_size')
+def modify_volume_element_geometry(volume_element, new_grid_size):    
+    
+    from scipy.ndimage import zoom
+
+    new_grid_size = np.array(new_grid_size)
+    zoom_factor = np.array(new_grid_size) / volume_element['grid_size']
+    new_elem_mat_idx = zoom(volume_element['element_material_idx'], zoom_factor, order=0)
+    
+    volume_element['grid_size'] = new_grid_size
+    volume_element['element_material_idx'] = new_elem_mat_idx
+
+    volume_element = validate_volume_element(volume_element)    
+    out = {'volume_element': volume_element}
+    return out    
+
+
 @func_mapper(task='visualise_volume_element', method='VTK')
 def visualise_volume_element(volume_element):
     path = Path('geom.vtr')